Re-applied all changes made after the pre_action work was added
