+++ doc/psidk/src/development/services/rust-service/testing.md
@@ -43,7 +43,9 @@ mod service {
     }
 }
 
-// packages takes a comma-delimited list and references the package.metadata.psibase.package-name (case-sensitive) in Cargo.toml. The test chain will be initialized with default packages + services listed here.
+/* packages takes a comma-delimited list and references the
+ * package.metadata.psibase.package-name (case-sensitive) in Cargo.toml).
+ * The test chain will be initialized with default packages + services listed here. */
 #[psibase::test_case(packages("example"))]
 fn test_arith(chain: psibase::Chain) -> Result<(), psibase::Error> {
     // Verify the actions work as expected.
