+++ rust/psibase_macros/psibase-macros-lib/src/tests/basics.rs
@@ -1,82 +1,6 @@
 #[cfg(test)]
 mod tests {
-    // use crate::identity_macro_impl;
     use crate::service_macro::service_macro_impl;
-
-    /*
-    #[test]
-    fn test_identity_macro_impl() {
-        use crate::tests::helpers::assert_tokens_eq;
-        use quote::quote;
-
-        let attr = quote! {
-            #[allow(non_snake_case)]
-        };
-        let before = quote! {
-            mod service {
-                #[action]
-                fn add(a: i32, b: i32) -> i32 {
-                    a + b
-                }
-
-                #[action]
-                fn multiply(a: i32, b: i32) -> i32 {
-                    a * b
-                }
-            }
-        };
-        let expected = quote! {
-            mod service {
-                #[action]
-                fn add(a: i32, b: i32) -> i32 {
-                    a + b
-                }
-
-                #[action]
-                fn multiply(a: i32, b: i32) -> i32 {
-                    a * b
-                }
-            }
-        };
-
-        let after = identity_macro_impl(attr, before);
-        assert_tokens_eq(&expected, &after);
-    }
-    */
-
-    /*
-    #[test]
-    fn test_basic_service_structure() {
-        use crate::tests::helpers::assert_tokens_eq;
-        use quote::quote;
-
-        let attr = quote! {};
-        let before = quote! {
-            mod service {
-                #[action]
-                fn add(a: i32, b: i32) -> i32 {
-                    a + b
-                }
-                #[event(history)]
-                fn something() {
-
-                }
-            }
-        };
-        let expected = quote! {
-            mod service {
-                #[action]
-                fn add(a: i32, b: i32) -> i32 {
-                    a + b
-                }
-            }
-        };
-
-        service_macro_impl(attr, before);
-        // println!("{:#?}", after);
-        // assert_tokens_eq(&expected, &after);
-    }
-    */
 
     #[test]
     #[should_panic]
