+++ rust/psibase_macros/psibase-macros-lib/src/tests/mod.rs
@@ -1,2 +1,1 @@
 pub mod basics;
-// pub mod helpers;
